+++ language/en/info_acp_hangman.php
@@ -1,7 +1,7 @@
 <?php
 /**
 *
-* @package Hangman v0.4.0
+* @package Hangman v0.5.0
 * @copyright (c) 2021 - 2022 Mike-on-Tour
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
@@ -35,9 +35,10 @@ if (empty($lang) || !is_array($lang))
 $lang = array_merge($lang, [
 	'ACP_MOT_HANGMAN'							=> 'Hangman',
 	'ACP_MOT_HANGMAN_SETTINGS'					=> 'Settings',
+	// General settings
 	'ACP_MOT_HANGMAN_EXT_SETTINGS'				=> 'Extension settings',
 	'ACP_MOT_HANGMAN_AUTODELETE'				=> 'Delete used terms?',
-	'ACP_MOT_HANGMAN_AUTODELETE_EXP'			=> 'If you select ´Yes´ then a term used in a game will be deleted from the data base after this game is finished.',
+	'ACP_MOT_HANGMAN_AUTODELETE_EXP'			=> 'If you select ´Yes´ then a term used in a game will be deleted from the database after this game is finished.',
 	'ACP_MOT_HANGMAN_CATEGORY_ENABLE'			=> 'Enable categories?',
 	'ACP_MOT_HANGMAN_CATEGORY_ENABLE_EXP'		=> 'If you select ´Yes´ an additional text field will be displayed while creating a search term. In this field you
 													can enter a category name to which this term belongs.<br>
@@ -45,13 +46,27 @@ $lang = array_merge($lang, [
 	'ACP_MOT_HANGMAN_CATEGORY_ENFORCE'			=> 'Enforce catgory input?',
 	'ACP_MOT_HANGMAN_CATEGORY_ENFORCE_EXP'		=> 'If you select ´Yes´ then the input of a category will be enforced which means that the input form can not be submitted
 													without an input in the category input field. This setting is only applicable if ´Enable categories´ is enabled.',
-	'ACP_MOT_HANGMAN_EVADE_ENABLE'				=> 'Count a game as lost if game is left?',
-	'ACP_MOT_HANGMAN_EVADE_ENABLE_EXP'			=> 'If you select ´Yes´ players leaving a running game (e.g. by klicking a link or refreshing the page) will be
-													scored the ´Points if losing´.<br>
-													If this option is enabled players will be asked via a popup window whether they really want to leave the game
-													page. If they affirm this action they will be scored (punished by) the ´Points if losing´. According to the
-													setting ´Delete used terms´ the search term used in this game may be deleted from the database.<br>
-													If this option is disabled the player wil be relayed to the selected page without further action.',
+	'ACP_MOT_HANGMAN_SHOW_TERM'					=> 'Display the term when closing a lost game',
+	'ACP_MOT_HANGMAN_SHOW_TERM_EXP'				=> 'if activated the term will be displayed in clear with the message about the lost game.',
+	'ACP_MOT_HANGMAN_ENFORCE_TERM'				=> 'Block players with an insufficient number of term input points',
+	'ACP_MOT_HANGMAN_ENFORCE_TERM_EXP'			=> 'If you select ´Yes´ players with a game points to term input points ratio exceeding the
+													set limit will be prompted to enter a new term prior to playing. This is a means to train
+													players neclecting the input of new terms.<br>
+													You can select the game points to term input points ratio in the next setting which will
+													be displayed after selecting ´Yes´.',
+	'ACP_MOT_HANGMAN_ENFORCE_TERM_RATIO'		=> 'Ratio of game points to term input points',
+	'ACP_MOT_HANGMAN_ENFORCE_TERM_RATIO_EXP'	=> 'Enabling this setting is primarily a means to block members playing Hangman if they entered an insufficient
+													number of new terms in relation to their winning games; it is therefore recommended to set this ratio
+													not too low. E.g. the default value of 40 means that a player with 40 term input points and more than
+													1,600 game points is blocked from playing until the ratio drops below 40 by entering a number of terms
+													to raise the term input points (the player will see a message stating that instead of the game panel).<br>
+													You should study the highscore table prior to initially setting this value in order to identify those players
+													with the greatest ratio, later you can lower this value. Your long-term goal should be to reach a ratio which
+													corresponds to the ratio of the mean winning game points and the points for entering a new term.<br>
+													As an example this means that with an average 25 game points and 8 term input points the exact ratio
+													would be 3.125 or 4 to bring up a round figure. With a ratio of 4 a player must enter a new term for
+													(nearly) every won game which seems to be a healthy figure if used terms are deleted. You can select
+													a higher value if used terms remain in the database.',
 	'ACP_MOT_HANGMAN_ROWS_PER_PAGE'				=> 'Rows per table page',
 	'ACP_MOT_HANGMAN_ROWS_PER_PAGE_EXP'			=> 'Choose the number of rows to be displayed per table page',
 	'ACP_MOT_HANGMAN_GAME_SETTINGS'				=> 'Game settings',
@@ -72,6 +87,14 @@ $lang = array_merge($lang, [
 	'ACP_MOT_HANGMAN_POINTS_WORD'				=> 'Points for a new quote',
 	'ACP_MOT_HANGMAN_POINTS_WORD_EXP'			=> 'Number of points to gain by entering a term or quote into the database<br>
 													(integer greater than or equal to zero)',
+	'ACP_MOT_HANGMAN_EVADE_ENABLE'				=> 'Count a game as lost if game is left?',
+	'ACP_MOT_HANGMAN_EVADE_ENABLE_EXP'			=> 'If you select ´Yes´ players leaving a running game (e.g. by klicking a link or refreshing the page) will be
+													scored the ´Points if losing´.<br>
+													If this option is enabled players will be asked via a popup window whether they really want to leave the game
+													page. If they affirm this action they will be scored (punished by) the ´Points if losing´. According to the
+													setting ´Delete used terms´ the search term used in this game may be deleted from the database.<br>
+													If this option is disabled the player wil be relayed to the selected page without further action.',
+	//Term settings
 	'ACP_MOT_HANGMAN_INPUT_SETTINGS'			=> 'Term input settings',
 	'ACP_MOT_HANGMAN_INPUT_SETTINGS_EXP'		=> 'Here you can configure the settings for the input of search terms.',
 	'ACP_MOT_HANGMAN_TERM_LENGTH'				=> 'Minimum length of the term',
@@ -110,6 +133,14 @@ $lang = array_merge($lang, [
 		2	=> 'Inserted %1$d data sets from table „<i>%2$s</i>“.',
 	],
 	'ACP_MOT_HANGMAN_TABLE_NO_IMPORT'			=> 'The table „<i>%1$s</i>“ does not contain any valid data.',
+	'ACP_MOT_HANGMAN_DATA_EXPORT'				=> 'Export data',
+	'ACP_MOT_HANGMAN_EXPORT_TERMS'				=> 'Export terms from table',
+	'ACP_MOT_HANGMAN_EXPORT_TERMS_EXP'			=> 'Here you can export the terms from the “_mot_hangman_words” table into a XML file.<br>
+													Clicking the button to the right creates a XML file with a pre-defined name and offers it to you for
+													download and local storing.',
+	'ACP_MOT_HANGMAN_SUBMIT_EXPORT'				=> 'Export data',
+	'ACP_MOT_HANGMAN_TABLE_NO_EXPORT'			=> 'Data export failed, unable to create file.',
+	// Misc
 	'ACP_MOT_HANGMAN_RESET_HIGHSCORE'			=> 'Reset Highscore',
 	'ACP_MOT_HANGMAN_RESET_HIGHSCORE_CAUTION'	=> '<strong>CAUTION:</strong> This action is irreversible!',
 	'ACP_MOT_HANGMAN_SCORE_TABLE_CLEARED'		=> '<strong>Hangman Highscore table cleared</strong>',